POC_UsingQdrant_Chatbot/requirements.txt
openai
python-dotenv
qdrant-client
langchain
langchain-openai
langchain-qdrant
langchain-community
langchain-tavily
python-docx
unstructured
pypdf
pdf2image
python-magic